PA4/requirements.txt
pyzmq
netifaces
requests
protobuf
influxdb3-python